+++ src/Braintree/IPaymentMethodGateway.cs
@@ -8,5 +8,7 @@ namespace Braintree
         Result<PaymentMethod> Delete(string token);
         PaymentMethod Find(string token);
         Result<PaymentMethod> Update(string token, PaymentMethodRequest request);
+        Result<PaymentMethodNonce> Grant(string token, PaymentMethodGrantRequest request);
+        Result<PaymentMethod> Revoke(string token);
     }
 }